Move formatter to correct dir
